==> 06.summary_stats.Rmd <==
---
title: "Summary Statistics"
author: "Andie Creel"
date: "2023-06-26"
output: pdf_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE)
knitr::opts_chunk$set(message  = FALSE)

rm(list = ls())
options(scipen = 999)
# devtools::install_github("UrbanInstitute/urbnmapr")
library(urbnmapr) # For map
library(dplyr)
library(tidyr)
library(stringr)
library(vroom)
library(mlogit)
library(stargazer)
library(lubridate)
library(kableExtra)
# library(choroplethr) #https://github.com/arilamstein/choroplethr
library(choroplethrMaps)
data("county.regions")
library(ggplot2)
library(ggpubr)
library(gridExtra)
```

# How often is indoor vs outdoor leisure is choosen

Assuming each choice occasion is a day, and an individual can choose indoor or outdoor leisure either at home or away from home. An individual can also choose no leisure. 

Choices have been multiplied by their sampling weight.

```{r}
# -----------------------------------------------------------------------------
# read in R memory built in 05.R_data_objs
# -----------------------------------------------------------------------------
load("05.RUM_data_objs.RData")


# -----------------------------------------------------------------------------
# Read in data 
# -----------------------------------------------------------------------------

# need replication weights 
myWts <- vroom("clean_data/2.demographics_ALL.csv") %>%
  select(caseid, year, wt)

myWorking <- myRUM_df %>%
  mutate(choice = if_else(number_activities > 0, 1, 0)) %>% # switching to extensive margin 
  left_join(myWts, by = c("year", "caseid")) %>% # merge in weights 
  mutate(choice_wt = choice*wt) # weighted choice ( whether you choose it times weight)

# -----------------------------------------------------------------------------
# How often does someone choose each thing (weights accounted for )
# -----------------------------------------------------------------------------

# function that takes as df and returns % of days of each activity is choosen 
getPercTable <- function(df){

  # divided by 4 because it's long by the 4 activities, multiplied by weight 
  n <- sum(df$wt) /4
  
  sumTable <- data.frame(activity = c("indoor_home", "indoor_away", "outdoor_away", "outdoor_home", "obs"), percent_days_choosen = NA)
  
  sumTable$percent_days_choosen[1] <- as.numeric(sum(df[df$activity == "indoor_home",]$choice_wt)/n)
  sumTable$percent_days_choosen[2] <- as.numeric(sum(df[df$activity == "indoor_away",]$choice_wt)/n)
  sumTable$percent_days_choosen[3] <- as.numeric(sum(df[df$activity == "outdoor_home",]$choice_wt)/n)
  sumTable$percent_days_choosen[4] <- as.numeric(sum(df[df$activity == "outdoor_away",]$choice_wt)/n)
  sumTable$percent_days_choosen[5] <- nrow(df)/5
  
  sumTable <- sumTable %>% 
    mutate(percent_days_choosen = round(percent_days_choosen *100, 1)) 
  
  sumTable
}

getPercTable(myWorking)

```

## by season

```{r}
# -----------------------------------------------------------------------------
# Grouping by season 
# -----------------------------------------------------------------------------

# my season function 
seasonTables <- lapply(X = c("spring", "summer", "fall", "winter"), function(x){
  myWorking %>%
    filter(season == x) %>%
    getPercTable() %>% 
    rename(!!sym(x) := percent_days_choosen) %>% 
    select(!!sym(x) )
  })

seasonTab <- bind_cols(activity = c("indoor home", "indoor away", "outdoor away", "outdoor home", "n"), 
          seasonTables) %>% 
  mutate(across(where(is.numeric), ~ round(., digits = 0))) %>% 
  mutate(across(where(is.numeric), ~ format(., big.mark = ",")))

seasonTab

# code for camp resources presentation 

latex_code <- kable(seasonTab, format = "latex", booktabs = TRUE, caption = "Percent of Americans Participating by Season")
cat(latex_code)

```

# Comparing full sample to my sample with weather 

When I run the RUM model, I drop people who do "no leisure" because they're less than 1% of the population. 

ATTN: Weather is an independent variable but only have if for half of observations so comparing full set to set I use for RUM model. I need to see if I can merge county weather on MSA's, I think that's why I'm loosing so many observations. 

Most of the data is dropped due to missing weather, some is dropped due to not having income.

```{r}
# -----------------------------------------------------------------------------
# get some key averages
# -----------------------------------------------------------------------------

getKeyAvgs <- function(df){

  # full data set 
  sumTable <- data.frame(metric = c("income", "% white", "min travel for leisure away", "min travl for rec away", "n"),
                         average = NA)
  
  mySum_df_indoor <- df %>%
    select(caseid, year, fam_inc_mid, travel_time_total_state, activity, race) %>%
    filter(activity == "indoor_away") %>%
    distinct()
  
  mySum_df_outdoor <- df %>%
    select(caseid, year, fam_inc_mid, travel_time_total_state, activity, race) %>%
    filter(activity == "outdoor_away") %>%
    distinct()
  
  sumTable$average[1] <- mean(mySum_df_indoor$fam_inc_mid, na.rm = T)
  sumTable$average[2] <- as.numeric(round(sum(mySum_df_indoor$race == "white_noH") / nrow(mySum_df_indoor) *100, 1))
  sumTable$average[3] <- mean(mySum_df_indoor$travel_time_total_state)
  sumTable$average[4] <- mean(mySum_df_outdoor$travel_time_total_state)
  sumTable$average[5] <- as.integer(nrow(mySum_df_indoor))
  
  sumTable
}

df_1 <-  myWorking %>%
  filter(activity != "no_leisure") %>%
  getKeyAvgs() %>%
  mutate(average_full_dataset = average) %>%
  select(average_full_dataset)

df_2 <- getKeyAvgs(myRUM_df) %>%
  mutate(average_RUM_dataset = average) %>%
  select(average_RUM_dataset)

bind_cols(metric = c("income", "% white", "min travel for leisure away", "min travl for rec away", "n"), df_1, df_2)

```

Seems like currently I'm dropping white poor people (which makes me think I'm dropping rural people). 

# Map of counties I have data for

```{r}
myCounties <- myRUM_df %>%
  # filter(!is.na(tmmx)) %>% 
  select(county) %>%
  distinct() %>%
  mutate(value = 1)
  

counties_sf_og <- get_urbn_map(map = "counties", sf = TRUE)
counties_sf <- get_urbn_map(map = "counties", sf = TRUE)

counties_sf <- counties_sf_og %>%
  mutate(county_fips = as.numeric(county_fips))

myMap <- left_join(counties_sf, myCounties, by =c("county_fips" = "county"))

myMap %>%
  ggplot() +
  geom_sf(mapping = aes(fill = value), color = NA) +
  geom_sf(data = counties_sf_og, fill = NA, color = "black", size = 0.25) +
  coord_sf(datum = NA) +   
  scale_fill_gradient(na.value="white") +
  theme_bw() + theme(legend.position="bottom", panel.border = element_blank()) +
  ggtitle("Counties that I currrently have data for")


```

There isn't a way to fix this, half of people dont say the MSA or the county they're living in so we only know what state. 


```{r}
# income not included and temp
reg1 <- mlogit(choice ~ travel_time_total_state, 
               myRUM_idx, 
               reflevel = "indoor_home") 

# temp included 
reg2 <- mlogit(choice ~ travel_time_total_state | tmmx, 
               myRUM_idx, 
               reflevel = "indoor_home")

stargazer(reg1, reg2, 
          title = "preferred model (entire year)",
          type = "text")
```




# Avg min doing activity 

```{r}
# -----------------------------------------------------------------------------
# read in data
# -----------------------------------------------------------------------------

myData_og <- vroom("clean_data/2.ATUS_wide_og.csv")
myDems <- vroom("clean_data/3.demographics_cleaned.csv")

myData <- myData_og %>%
  group_by(date, caseid) %>%
  mutate(total_outdoor_away = sum(time_recreating_away_temp, na.rm = T)) %>%
  mutate(total_outdoor_home = sum(time_recreating_home_temp, na.rm = T)) %>%
  mutate(total_indoor_home = sum(time_leisure_home_temp, na.rm = T)) %>%
  mutate(total_indoor_away = sum(time_leisure_away_temp, na.rm = T)) %>%
  select(year, caseid, date, starts_with("total"), wt06, wt20) %>%
  distinct() %>%
  mutate(wt = if_else(year != 2020, wt06, wt20)) %>%
  select(-wt06, -wt20)     


# -----------------------------------------------------------------------------
# Calculating average time per year
# In zotero: “American Time Use Survey User’s Guide,” 2022, 125.
# -----------------------------------------------------------------------------

myData_graphs <- myData %>%
  group_by(year) %>%
  mutate(outdoor_away = sum(wt*total_outdoor_away)/sum(wt)) %>%
  mutate(outdoor_home = sum(wt*total_outdoor_home)/sum(wt)) %>%
  mutate(indoor_away = sum(wt*total_indoor_away)/sum(wt)) %>%
  mutate(indoor_home = sum(wt*total_indoor_home)/sum(wt)) %>%
  select(year, 
         outdoor_away, outdoor_home,
         indoor_away, indoor_home
         ) %>%
  distinct()
           
myGraph_long <- myData_graphs %>%
  pivot_longer(
    cols = c(outdoor_away,outdoor_home, indoor_away,  indoor_home), # Select the columns to pivot
    names_to = "activity",       # Name of the new 'activity' column
    values_to = "minutes"          # Name of the new 'value' column
  )


# -----------------------------------------------------------------------------
# Leisure graph
# -----------------------------------------------------------------------------
myGraph_indoor <- myGraph_long %>%
  filter(str_detect(activity, "^indoor"))

# Plotting the data using ggplot
plot_indoor_1 <- ggplot(myGraph_indoor, aes(x = year, y = minutes, color = activity)) +
  geom_line() +
  labs(x = "Year", y = "Minutes") +
  scale_color_discrete(name = "activity") +
  ggtitle("Avg min on indoor per day") +
  theme_bw()

```
 

```{r}
# -----------------------------------------------------------------------------
# Recreations graph
# -----------------------------------------------------------------------------
myGraph_outdoor<- myGraph_long %>%
  filter(str_detect(activity, "^outdoor"))

# Plotting the data using ggplot
plot_outdoor_1 <- ggplot(myGraph_outdoor, aes(x = year, y = minutes, color = activity)) +
  geom_line() +
  labs(x = "Year", y = "Minutes") +
  scale_color_discrete(name = "activity") +
  ggtitle("Avg min on outdoor per day") +
  theme_bw()

```


```{r}
# -----------------------------------------------------------------------------
# creating cyclical maps
# -----------------------------------------------------------------------------

myData_seasons <- myData_og %>%
  mutate(year_q = year+(quarter-1)*.25) %>%
  group_by(date, caseid) %>%
  mutate(total_outdoor_away = sum(time_recreating_away_temp, na.rm = T)) %>%
  mutate(total_outdoor_home = sum(time_recreating_home_temp, na.rm = T)) %>%
  mutate(total_indoor_home = sum(time_leisure_home_temp, na.rm = T)) %>%
  mutate(total_indoor_away = sum(time_leisure_away_temp, na.rm = T)) %>%
  select(year, caseid, date, starts_with("total"), wt06, wt20, year_q) %>%
  distinct() %>%
  mutate(wt = if_else(year != 2020, wt06, wt20)) %>%
  select(-wt06, -wt20) %>%   
  group_by(year_q) %>%
  mutate(outdoor_away = sum(wt*total_outdoor_away)/sum(wt)) %>%
  mutate(outdoor_home = sum(wt*total_outdoor_home)/sum(wt)) %>%
  mutate(indoor_away = sum(wt*total_indoor_away)/sum(wt)) %>%
  mutate(indoor_home = sum(wt*total_indoor_home)/sum(wt)) %>%
  select(year, year_q,
         outdoor_away, outdoor_home,
         indoor_away, indoor_home
         ) %>%
  distinct()

mySeasons_long <- myData_seasons %>%
  pivot_longer(
    cols = c(outdoor_away,outdoor_home, indoor_away,  indoor_home), # Select the columns to pivot
    names_to = "activity",      
    values_to = "minutes"          
  )

# -----------------------------------------------------------------------------
# Leisure graph
# -----------------------------------------------------------------------------
mySeasons_indoor <- mySeasons_long %>%
  filter(str_detect(activity, "^indoor"))

# Plotting the data using ggplot
plot_indoor_2 <- ggplot(mySeasons_indoor, aes(x = year_q, y = minutes, color = activity)) +
  geom_line() +
  labs(x = "Year", y = "Minutes") +
  scale_color_discrete(name = "activity") +
  ggtitle("Avg min on indoor per day") +
  theme_bw()

```



```{r}
# -----------------------------------------------------------------------------
# recreation graph
# -----------------------------------------------------------------------------
mySeasons_outdoor <- mySeasons_long %>%
  filter(str_detect(activity, "^outdoor"))

# Plotting the data using ggplot
plot_outdoor_2 <-  ggplot(mySeasons_outdoor, aes(x = year_q, y = minutes, color = activity)) +
  geom_line() +
  labs(x = "Year", y = "Minutes") +
  scale_color_discrete(name = "activity") +
  ggtitle("Avg min on outdoor per day") +
  theme_bw()

grid.arrange(plot_indoor_1, plot_outdoor_1, plot_indoor_2, plot_outdoor_2, 
             nrow = 2, ncol = 2)

```

Both indoor activities have held steady over the study period, with the exception of during COVID Outdoor leisure is also fairly steady with a potential spike during COVID. 

Cyclical trend on second row. Indoor activities away goes up in summer months and down in winter. Opposite is true for indoor activities at home. 

Way easier to see the cyclical trend with outdoor recreation.

# By race

```{r}
# -----------------------------------------------------------------------------
# what races am i including
# -----------------------------------------------------------------------------
myDems_race <- myDems %>%
  select(caseid, race) 

race_counts <- table(myDems_race$race)/nrow(myDems_race)

# Include: white_noH, black, hispanic (combined), asian, other

myDems_race <- myDems_race %>%
  mutate(race_1 = race) %>%
  mutate(race_1 = if_else(race == "hispanic_black" | race == "hispanic_white", 
         "hispanic", race_1)) %>%
  mutate(race_1 = if_else(race == "islander" | race == "native" | race == "other" | race == "white-asian" | race == "white-black" | race == "white-native", 
         "other", race_1)) 

race_counts <- table(myDems_race$race_1)/nrow(myDems_race)

```


```{r}

myRace <- left_join(myData, myDems_race, by = "caseid")

myRace_graphs <- myRace %>%
  group_by(year, race_1) %>%
  mutate(outdoor_away = sum(wt*total_outdoor_away)/sum(wt)) %>%
  mutate(outdoor_home = sum(wt*total_outdoor_home)/sum(wt)) %>%
  mutate(indoor_away = sum(wt*total_indoor_away)/sum(wt)) %>%
  mutate(indoor_home = sum(wt*total_indoor_home)/sum(wt)) %>%
  select(year, 
         outdoor_away, outdoor_home,
         indoor_away, indoor_home,
         race_1) %>%
  distinct()
           
myRace_long <- myRace_graphs %>%
  pivot_longer(
    cols = c(outdoor_away, outdoor_home, indoor_away,  indoor_home), # Select the columns to pivot
    names_to = "activity",       # Name of the new 'activity' column
    values_to = "minutes"          # Name of the new 'value' column
  ) %>%
  filter(!str_detect(race_1, "^other")) %>%  # dropping the "other" race
  filter(year < 2020)

# Create a named vector of colors to map to specific race categories
custom_colors <- c("asian" = "darkgreen", "black" = "darkblue", 
                   "hispanic" = "darkred", "white_noH" = "darkorange")

custom_names <- c("asian" = "Asian", "black" = "Black", 
                   "hispanic" = "Hispanic", "white_noH" = "White (non-hisp)")

# -----------------------------------------------------------------------------
# Leisure home graph
# -----------------------------------------------------------------------------
myRace_indoor_home <- myRace_long %>%
  filter(str_detect(activity, "^indoor_home")) 

# Plotting the data using ggplot
plot_race_1 <- ggplot(myRace_indoor_home, aes(x = year, y = minutes, color = race_1)) +
  geom_point(alpha = 0.6, size = 3) +
  geom_smooth(method = "lm", se = F, size = 1.5)+
  labs(x = "Year", y = "Minutes") +
  scale_color_manual(values = custom_colors, 
                     labels = custom_names) +
  ggtitle("Indoor at Home") +
  theme_bw()+ 
  labs(color = "Race")+ 
  theme(text = element_text(size=20))

# -----------------------------------------------------------------------------
# Leisure away graph
# -----------------------------------------------------------------------------
myRace_indoor_away <- myRace_long %>%
  filter(str_detect(activity, "^indoor_away"))

# Plotting the data using ggplot
plot_race_2 <-ggplot(myRace_indoor_away, aes(x = year, y = minutes, color = race_1)) +
  geom_point(alpha = 0.6, size = 3) +
  geom_smooth(method = "lm", se = F, size = 1.5)+
  labs(x = "Year", y = "Minutes") +
  scale_color_manual(values = custom_colors, 
                     labels = custom_names) +
  ggtitle("Indoor Away from Home") +
  theme_bw()+ 
  labs(color = "Race")+ 
  theme(text = element_text(size=20))

# -----------------------------------------------------------------------------
# rec home graph
# -----------------------------------------------------------------------------
myRace_outdoor_home <- myRace_long %>%
  filter(str_detect(activity, "^outdoor_home")) 

# Plotting the data using ggplot
plot_race_3 <-ggplot(myRace_outdoor_home, aes(x = year, y = minutes, color = race_1)) +
  geom_point(alpha = 0.6, size = 3) +
  geom_smooth(method = "lm", se = F, size = 1.5)+
  labs(x = "Year", y = "Minutes") +
  scale_color_manual(values = custom_colors, 
                     labels = custom_names) +
  ggtitle("Outdoor at Home") +
  theme_bw()+ 
  labs(color = "Race")+ 
  theme(text = element_text(size=20))

# -----------------------------------------------------------------------------
# Leisure away graph
# -----------------------------------------------------------------------------
myRace_outdoor_away <- myRace_long %>%
  filter(str_detect(activity, "^outdoor_away"))

# Plotting the data using ggplot
plot_race_4 <-ggplot(myRace_outdoor_away, aes(x = year, y = minutes, color = race_1)) +
  geom_point(alpha = 0.6, size = 3) +
  geom_smooth(method = "lm", se = F, size = 1.5)+
  labs(x = "Year", y = "Minutes") +
  scale_color_manual(values = custom_colors, 
                     labels = custom_names) +
  ggtitle("Outdoor Away from Home") +
  theme_bw() + 
  labs(color = "Race")+ 
  theme(text = element_text(size=20))


# Combine the plots using ggarrange
combined_plot <- ggarrange(plot_race_1, plot_race_2, plot_race_3, plot_race_4, 
                           ncol = 2, nrow = 2, common.legend = TRUE, legend = "bottom")
combined_plot

ggsave(combined_plot, width = 13, height = 8, filename = "figures/minutes_per_day.jpeg")
```

Black people participate in much more leisure at home. 

Black and Asian people do less recreation at home, and black people participate in much less recreation in general.
