==> 07.elasticities_RUM_writeup.Rmd <==
---
title: "RUM Model: elasticities and temperature affects"
author: "Andie Creel"
date: "2023-07-06"
output: 
  pdf_document:
    number_sections: true
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE)
knitr::opts_chunk$set(message  = FALSE)
```

# Questions of interest: 

As summers heat up, how can individuals adapt the time of year they recreate outside? How can institutions like schools and work change so that people can adapt their recreation schedule. 
  
Note: all results are on the extensive margin (did you take a trip that day or not). 

From the 05.write_up.pdf I found that constructing the travel cost for no trips is fairly robust, regardless of if I'm using total travel time or average travel time per trip, as well as if I'm grouping by race or state. 

In this write up, I use **total travel time grouped by state** to get the cost when people don't take a trip. 


# RUM model (Annual)

Indirect utility:

$$V_{itc} = \alpha_c + \beta x_{ic} + \gamma_c w_t $$
Where $V$ is the indirect utility for person $i$ on choice occasion $t$ (day) for choice $c$. $\alpha$ is a choice specific constant. $\beta$ is a general coefficient for travel **time** $x$. $\gamma$ is a choice specific coefficient for weather on day $t$ at for the choice occasion $w$. 


```{r}
rm(list = ls())
options(scipen = 999)
library(dplyr)
library(tidyr)
library(stringr)
library(vroom)
library(mlogit)
library(stargazer)
library(lubridate)
library(ggplot2)
library(gridExtra)


# -----------------------------------------------------------------------------
# read in R memory built in R_data_objs
# -----------------------------------------------------------------------------
load("05.RUM_data_objs.RData")

# -----------------------------------------------------------------------------
# preferred reg specification (total travel time grouped by state)
# -----------------------------------------------------------------------------

# income not included
reg1.a <- mlogit(choice ~ travel_time_total_state |  tmmx, # formula
               myRUM_idx, #mlogit data object
               reflevel = "indoor_home", #reference level 
               ) # choices available (dropped no leisure)

# income included 
reg1.b <- mlogit(choice ~ travel_time_total_state | fam_inc_mid + tmmx, # formula
               myRUM_idx, 
               reflevel = "indoor_home") 

stargazer(reg1.a, reg1.b,
          title = "preferred model (entire year)",
          type = "text")

```
I'm currently going to work with the regression where income is not included because the estimates are robust. The numeraire is currently minutes. 

## Marginal Effects 

In multinomial logit models, the coefficients are NOT the marginal effects. 

Marginal effect equations: 
$$\frac{\partial P_{itl}}{\partial x_{itl}} = \beta P_{itl}(1 - P_{itl})$$

$$\frac{\partial P_{itl}}{\partial w_{t}} = P_{itl}(\gamma_l - \sum_c P_{itc} \gamma_c)$$

$$\frac{\partial P_{itl}}{\partial x_{itk}} = - \beta P_{itl}P_{itk}$$
I calculate the marginal effects as a relative change to a relative change (baseline is the average) so that marginal effects are an elasticity. 


### Elasticities for temperature 
```{r}
# -----------------------------------------------------------------------------
# preferred reg specification (total travel time grouped by state)
# -----------------------------------------------------------------------------
elasts.temp.1a <- effects(reg1.a, covariate = "tmmx", type = "rr")

elasts.temp.1a
```

A 1% increase in temperature leads: 

  - `r round(elasts.temp.1a["indoor_home"],2)`% change in probability of choosing indoor leisure at home 
  - `r round(elasts.temp.1a["indoor_away"],2)`% change in probability of choosing indoor leisure away from home 
  - `r round(elasts.temp.1a["outdoor_away"],2)`% change in probability of choosing outdoor leisure away from home
  - `r round(elasts.temp.1a["outdoor_home"],2)`% change in probability of choosing outdoor leisure at home 
  
Note: I find it reassuring that the outdoor leisure at home and away from home increase in a similar way. 

### Elasticities for travel time

```{r}
elasts.travel.1a<- effects(reg1.a, covariate = "travel_time_total_state", type = "rr")
round(elasts.travel.1a, 2)
```

A 1% increase in travel time to indoor away from home activities leads to: 

  - `r round(elasts.travel.1a["indoor_away", "indoor_home"], 2)`% change in probability of choosing indoor home, outdoor away and outdoor home leisure (IIA property)
  - `r round(elasts.travel.1a["indoor_away", "indoor_away"], 2)`% change in probability of choosing indoor away from leisure 
  
  
A 1% increase in travel time to outdoor away from home leisure activities leads to: 

  - `r round(elasts.travel.1a["outdoor_away", "indoor_home"], 3)`% change in probability of choosing indoor home, indoor away and outdoor home leisure (IIA property)
  - `r round(elasts.travel.1a["outdoor_away", "outdoor_away"], 2)`% change in probability of choosing outdoor away from home leisure 

## Travel cost calcs

```{r}
# -----------------------------------------------------------------------------
# getting OC of 1 degree temp increase (in minutes)
# https://cran.r-project.org/web/packages/mlogit/vignettes/c3.rum.html
# -----------------------------------------------------------------------------

getTravelCost_temp <- function(reg){
  coef(reg)[grep("tmmx", names(coef(reg)))] /coef(reg)["travel_time_total_state"] 
}

getTravelCost_temp(reg1.a)


```

For the average day of the year, people would travel an additional ~9 minutes further to get to outdoor recreation that is 1 degree warming and 3 minutes for indoor leisure. 



# By Season 

```{r}
# -----------------------------------------------------------------------------
# Season regressions
# -----------------------------------------------------------------------------

mySeasonReg <- function(i){

  myWorking_temp <- myRUM_idx %>%
    filter(season == i) 
  
  reg <- mlogit(choice ~ travel_time_total_state | tmmx, # formula
                         myWorking_temp, #mlogit data object
                         reflevel = "indoor_home") #reference level
  
}

seasonRegs <- lapply(c("spring","summer","fall","winter"), 
                      mySeasonReg)

stargazer(seasonRegs, 
          column.labels = c("spring","summer","fall","winter"),
          title = "The effect of temperature changes given the season.",
          type = "text")



```

## Elasticities (marginal effects)

### Temperature 

```{r}
# -----------------------------------------------------------------------------
# seasonal marginal effects
# -----------------------------------------------------------------------------

myEffects_temp <- function(x){
  effects(x, covariate = "tmmx", type = "rr")
}

season_temp_elasts <- lapply(seasonRegs, myEffects_temp) %>%
  bind_rows() %>%
  round(2)

season_temp_elasts <- bind_cols(season = c("spring","summer","fall","winter"), 
                                season_temp_elasts)
season_temp_elasts

```

### Travel Time 

```{r}

# -----------------------------------------------------------------------------
# seasonal marginal effects
# -----------------------------------------------------------------------------

myEffects_time <- function(x){
  effects(x, covariate = "travel_time_total_state", type = "rr")
}

season_time_elasts_list <- lapply(seasonRegs, myEffects_time) 

season_time_elasts <- bind_cols(season = c("spring","summer","fall","winter"), 
                                indoor_away = c(season_time_elasts_list[[1]]["indoor_away", "indoor_away"], 
                                                season_time_elasts_list[[2]]["indoor_away", "indoor_away"], 
                                                season_time_elasts_list[[3]]["indoor_away", "indoor_away"], 
                                                season_time_elasts_list[[4]]["indoor_away", "indoor_away"]),
                                outdoor_away = c(season_time_elasts_list[[1]]["outdoor_away", "outdoor_away"], 
                                                season_time_elasts_list[[2]]["outdoor_away", "outdoor_away"], 
                                                season_time_elasts_list[[3]]["outdoor_away", "outdoor_away"], 
                                                season_time_elasts_list[[4]]["outdoor_away", "outdoor_away"]))
season_time_elasts

```

Demand for away from home activites is the least elastic in the summer and the most elastic in the winter (for a 1% increase in travel time). However, all of these are less than 1% in magnitude. Responces in demand to changes in travel time if far more inelastic than temperature changes.


## Travel time cost by season

```{r}
# -----------------------------------------------------------------------------
# getting OC of 1 degree temp increase (in minutes) by season
# https://cran.r-project.org/web/packages/mlogit/vignettes/c3.rum.html
# -----------------------------------------------------------------------------

lapply(seasonRegs, getTravelCost_temp) %>%
  bind_rows() %>%
  bind_cols(season = c("spring","summer","fall","winter"),) %>%
  select(season, starts_with("tmmx"))


```

In the summer, people lose an avg of 8 - 13 minutes of welfare for a temperature increase from away from home activities.  

In other seasons, there are welfare gains for temp increases but they're smaller. 


## What are the average number of trips taken per season? 
This is on the *intensive* margin again, so counts all the activities participated in a day. 

```{r}
# -----------------------------------------------------------------------------
# get average number of total trips per season
# -----------------------------------------------------------------------------

# need replication weights 
myWts <- vroom("clean_data/2.demographics_ALL.csv") %>%
  mutate(wt = if_else(year == 2020, wt20, wt06)) %>%
  select(caseid, year, wt)

avgTripsSeason <- myRUM_df %>%
  left_join(myWts, by = c("year", "caseid")) %>% 
  group_by(year, activity, season) %>%
  summarize(total_days = sum(number_activities*wt)) %>% #intesive margin
  group_by(activity, season) %>%
  mutate(avg_trips = mean(total_days)/1000000) %>% 
  select(activity, season, avg_trips) %>% 
  distinct()
  

# Define the order of the activities for proper ordering on the x-axis
activity_order <- c("indoor_home", "outdoor_home", "indoor_away", "outdoor_away")

# -----------------------------------------------------------------------------
# Indoor activities 
# -----------------------------------------------------------------------------

avgTripsSeason_indoor <- avgTripsSeason %>% 
  filter(str_detect(activity, "^indoor"))

# Create the bar chart
plot_indoor <- ggplot(avgTripsSeason_indoor, aes(x = season, y = avg_trips, fill = activity)) +
  geom_bar(stat = "identity", position = "dodge") +
  scale_fill_manual(values = c("#E69F00", "#56B4E9")) +
  labs(x = "Season", y = "Average trips (millions)") +
  scale_x_discrete(limits = c("spring", "summer", "fall", "winter")) +
  theme_bw() +
  theme(legend.position = "top") +
  guides(fill = guide_legend(title = "Activity"))


# -----------------------------------------------------------------------------
# outdoor activities 
# -----------------------------------------------------------------------------
avgTripsSeason_outdoor <- avgTripsSeason %>% 
  filter(str_detect(activity, "^outdoor"))

# Create the bar chart
plot_outdoor <- ggplot(avgTripsSeason_outdoor, aes(x = season, y = avg_trips, fill = activity)) +
  geom_bar(stat = "identity", position = "dodge") +
  scale_fill_manual(values = c("#F0E442", "#009E73")) +
  labs(x = "Season", y = "Average trips (millions)") +
  scale_x_discrete(limits = c("spring", "summer", "fall", "winter")) +
  theme_bw() +
  theme(legend.position = "top") +
  guides(fill = guide_legend(title = "Activity"))


grid.arrange(plot_indoor, plot_outdoor,
             nrow = 1, ncol = 2)

```
Outdoor activities see a spike in the spring and summer, but indoor activities are relatively more steady through the year.


This is only our sample, not the broader sample (bc a lot of data dropped with temperature stuff).


## Have the elasticities trended through time? 

```{r}
# -----------------------------------------------------------------------------
# functions
# -----------------------------------------------------------------------------
getSeasonElasticity <- function(y, s) {
  
  myWorking_temp <- myRUM_idx %>%
    # filter(season == "summer") %>%
    filter(season == s) %>% 
    filter(year == y)
  
  reg <- mlogit(choice ~ travel_time_total_state | tmmx, # formula
                         myWorking_temp, #mlogit data object
                         reflevel = "indoor_home") #reference level
  
  effects(reg, covariate = "tmmx", type = "rr")
}


# -----------------------------------------------------------------------------
# getting graphs of seasonal elasticity through time
# -----------------------------------------------------------------------------

getSeasonTempGraph <- function(season){
  
  #run regs and turn into table
  season_temp_elasts_time <- lapply(2003:2021, getSeasonElasticity, s = season) %>% 
    bind_rows()
  
  # add in year
  season_temp_elasts_time <- bind_cols(year = 2003:2021, 
                                        season_temp_elasts_time)
  
  
  # pivot long for graph
  seasons_elasts_long <- season_temp_elasts_time %>%
    pivot_longer(
      cols = c(outdoor_away,outdoor_home, indoor_away,  indoor_home), # Select the columns to pivot
      names_to = "activity",      
      values_to = "temp_elasticity"          
    )
  
  # graph
   ggplot(seasons_elasts_long, aes(x = year, y = temp_elasticity, color = activity)) +
    geom_point() +
    geom_smooth(method = "lm", se = F) +
    labs(x = "Year", y = "Elasticity") +
    scale_color_discrete(name = "activity") +
    geom_hline(yintercept = 0) +
    ggtitle(season) +
    theme_bw()+ 
    theme(legend.position = "right") +
    guides(fill = guide_legend(title = "Activity"))
}


temp_elast_graphs <- lapply(c("spring","summer","fall","winter"), getSeasonTempGraph)

grid.arrange(temp_elast_graphs[[1]], temp_elast_graphs[[2]], 
             temp_elast_graphs[[3]], temp_elast_graphs[[4]],
             nrow = 2, ncol = 2,
             top = "Responce to  1% temp increase")





```
An increase summer temperature drives people home. This trend has only been exaggerated through time (potentially as the world has warmed through the last 20 years).
